move temporal tag routes outside the multimodal gate
