+++ fiftyone/multimodal/server/__init__.py
@@ -6,15 +6,21 @@ Server scaffolding for multimodal workflows.
 |
 """
 
+from typing import TYPE_CHECKING
+
 from .routes import (
     MultimodalRoutes,
     PlaybackPlanEndpoint,
     PROTOBUF_MEDIA_TYPE,
-    SampleTagsEndpoint,
     SceneInventoryEndpoint,
-    TagCountsEndpoint,
-    TagsEndpoint,
 )
+
+if TYPE_CHECKING:
+    from fiftyone.server.routes.temporal_tags import (
+        SampleTagsEndpoint,
+        TagCountsEndpoint,
+        TagsEndpoint,
+    )
 
 __all__ = [
     "MultimodalRoutes",
@@ -25,3 +31,21 @@ __all__ = [
     "TagCountsEndpoint",
     "TagsEndpoint",
 ]
+
+
+_TEMPORAL_TAG_EXPORTS = {
+    "SampleTagsEndpoint",
+    "TagCountsEndpoint",
+    "TagsEndpoint",
+}
+
+
+def __getattr__(name):
+    if name in _TEMPORAL_TAG_EXPORTS:
+        from fiftyone.server.routes import (  # pylint: disable=import-outside-toplevel
+            temporal_tags,
+        )
+
+        return getattr(temporal_tags, name)
+
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
