+++ fiftyone/server/routes/__init__.py
@@ -30,6 +30,7 @@ from .screenshot import Screenshot
 from .sort import Sort
 from .tag import Tag
 from .tagging import Tagging
+from .temporal_tags import TemporalTagRoutes
 from .values import Values
 from .video_labels import VideoLabelsIndex, VideoLabelsWindow
 
@@ -44,6 +45,7 @@ routes = (
     CameraRoutes
     + EmbeddingsRoutes
     + GroupsRoutes
+    + TemporalTagRoutes
     + multimodal_routes
     + OperatorRoutes
     + RuntimeAssetRoutes
